Python-Project-On-Diwali-Sales-Analysis
The Diwali festival, celebrated with great fervor across India and in many parts of the world, is not only a time for festivities but also a significant period for businesses, especially retail, to capitalize on increased consumer spending. This Python project aims to analyze Diwali sales data to uncover trends, patterns, and insights that can inform business strategies.